pyproject.toml
[tool.poetry]
name = "preprocessor"
version = "0.1.0"
description = ""
authors = ["Krishna <krishna.gopala@bcbsnc.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^6.0"
psycopg2-binary = "^2.9.1"
snowflake-connector-python = "^2.7.0"
boto3 = "^1.19.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
